feat(columns): created a generic toggleButton column
